refactor(fuse): convert write_ops.rs to directory module with deduped bin-publish helper
- crates/fuse/src/write_ops.rs -> write_ops/mod.rs behind preserved pub(crate) mod implementation facade
- Handler bodies split into implementation/{file_data,delete,mkdir,rename}.rs
- Shared ~50-line bin-publish tail deduped into publish_bin_entry_on_delete helper in delete.rs
- crate::write_ops::implementation::handle_* caller paths unchanged
- All 64 cipherbox-fuse tests pass
